Change type of level to short, WIP guild skill database

+++ Scripts/Gameplay/Social/SocialCharacterData.cs
@@ -5,7 +5,7 @@
         public string id;
         public string characterName;
         public int dataId;
-        public int level;
+        public short level;
         public int currentHp;
         public int maxHp;
         public int currentMp;
